Add optional sendData callback to ContactForm component interface. Defaults to legacy behavior
